LayoutTests/discovery/get-network-services-types-add-dev-expected.txt

Tests navigator.getNetworkServices() works in happy case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() called ok CDS callback.
PASS srvsCDS.length==0 is true
PASS navigator.getNetworkServices() called ok Cnn callback.
PASS srvsCnn.length==0 is true
PASS successfullyParsed is true

TEST COMPLETE
